requirements.txt
prettytable==2.1.0
nibabel==3.0.2
torch==1.5.0
torchvision==0.6.0
opencv_python==4.1.0.25
matplotlib==3.1.0
scikit_image==0.16.2
numpy==1.19.5
scipy==1.1.0
Pillow==8.2.0
PyYAML==5.4.1
skimage==0.0
tensorboardX==2.2
tifffile==2021.4.8
